Split pool to containers and bits
